Address code review feedback and fix build error
Code review fixes:
- Add explanatory comment for reinterpret_cast in ROCm dlsym usage
- Add explicit cast to int for ImGui::PushID (size_t to int)
- Update completed-features.md to reflect Phase 7 completion
- Use emplace_back consistently in LinuxGPUProbe
- Update CONTRIBUTING.md to use clang-22 (matching requirements)

Build fix:
- Add missing <format> header to ProcessDetailsPanel.cpp
  (needed for std::format usage in GPU rendering code)

+++ src/App/Panels/SystemMetricsPanel.cpp
@@ -1464,7 +1464,7 @@ void SystemMetricsPanel::renderGpuSection()
         // GPU header with collapsible section
         const std::string headerLabel = std::format("{} {} [{}]", ICON_FA_MICROCHIP, gpuName, isIntegrated ? "Integrated" : "Discrete");
 
-        ImGui::PushID(gpuIdx);
+        ImGui::PushID(static_cast<int>(gpuIdx)); // gpuIdx is a small index; explicit narrowing to match ImGui API
         const bool expanded = ImGui::CollapsingHeader(headerLabel.c_str(), ImGuiTreeNodeFlags_DefaultOpen);
         ImGui::PopID();
